Extract Compose UI to compose-ui module
