+++ examples/todoapp/android/src/main/java/example/todo/android/MainActivity.kt
@@ -13,7 +13,7 @@ import com.arkivanov.mvikotlin.main.store.DefaultStoreFactory
 import com.arkivanov.mvikotlin.timetravel.store.TimeTravelStoreFactory
 import example.todo.common.database.TodoDatabaseDriver
 import example.todo.common.root.TodoRoot
-import example.todo.common.root.ui.TodoRootContent
+import example.todo.common.ui.TodoRootContent
 import example.todo.database.TodoDatabase
 
 class MainActivity : AppCompatActivity() {
