+++ examples/todoapp/buildSrc/src/main/kotlin/multiplatform-compose-setup.gradle.kts
@@ -1,29 +1,39 @@
 import org.jetbrains.compose.compose
 
 plugins {
+    id("com.android.library")
     id("kotlin-multiplatform")
     id("org.jetbrains.compose")
 }
 
 kotlin {
+    jvm("desktop")
+    android()
+
     sourceSets {
         named("commonMain") {
             dependencies {
-                api(compose.runtime)
-                api(compose.foundation)
-                api(compose.material)
+                implementation(compose.runtime)
+                implementation(compose.foundation)
+                implementation(compose.material)
             }
         }
+
         named("androidMain") {
             dependencies {
-                api("androidx.appcompat:appcompat:1.1.0")
-                api("androidx.core:core-ktx:1.3.1")
+                implementation("androidx.appcompat:appcompat:1.1.0")
+                implementation("androidx.core:core-ktx:1.3.1")
             }
         }
+
         named("desktopMain") {
             dependencies {
-                api(compose.desktop.common)
+                implementation(compose.desktop.common)
             }
         }
+    }
+
+    tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+        kotlinOptions.jvmTarget = "1.8"
     }
 }
