+++ examples/todoapp/buildSrc/src/main/kotlin/multiplatform-setup.gradle.kts
@@ -8,15 +8,6 @@ kotlin {
     android()
 
     sourceSets {
-        named("commonMain") {
-            dependencies {
-                implementation(Deps.ArkIvanov.MVIKotlin.mvikotlin)
-                implementation(Deps.ArkIvanov.MVIKotlin.mvikotlinExtensionsReaktive)
-                implementation(Deps.ArkIvanov.Decompose.decompose)
-                implementation(Deps.Badoo.Reaktive.reaktive)
-            }
-        }
-
         named("commonTest") {
             dependencies {
                 implementation(Deps.JetBrains.Kotlin.testCommon)
@@ -38,26 +29,5 @@ kotlin {
 
     tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
         kotlinOptions.jvmTarget = "1.8"
-    }
-}
-
-android {
-    compileSdkVersion(30)
-
-    defaultConfig {
-        minSdkVersion(23)
-        targetSdkVersion(30)
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-
-    sourceSets {
-        named("main") {
-            manifest.srcFile("src/androidMain/AndroidManifest.xml")
-            res.srcDirs("src/androidMain/res")
-        }
     }
 }
