+++ examples/todoapp/common/compose-ui/build.gradle.kts
@@ -1,21 +1,17 @@
 plugins {
-    id("multiplatform-setup")
+    id("multiplatform-compose-setup")
     id("android-setup")
-    id("kotlin-parcelize")
 }
 
 kotlin {
     sourceSets {
         named("commonMain") {
             dependencies {
-                implementation(project(":common:utils"))
-                implementation(project(":common:database"))
                 implementation(project(":common:main"))
                 implementation(project(":common:edit"))
-                implementation(Deps.ArkIvanov.MVIKotlin.mvikotlin)
+                implementation(project(":common:root"))
                 implementation(Deps.ArkIvanov.Decompose.decompose)
                 implementation(Deps.ArkIvanov.Decompose.extensionsCompose)
-                implementation(Deps.Badoo.Reaktive.reaktive)
             }
         }
     }
