+++ examples/todoapp/common/compose-ui/src/commonMain/kotlin/example/todo/common/ui/Crossfade.kt
@@ -1,4 +1,4 @@
-package example.todo.common.utils
+package example.todo.common.ui
 
 import androidx.compose.runtime.Composable
