+++ examples/todoapp/common/compose-ui/src/commonMain/kotlin/example/todo/common/ui/TodoEditUi.kt
@@ -1,4 +1,4 @@
-package example.todo.common.edit.ui
+package example.todo.common.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
