+++ examples/todoapp/common/compose-ui/src/commonMain/kotlin/example/todo/common/ui/TodoRootUi.kt
@@ -1,12 +1,9 @@
-package example.todo.common.root.ui
+package example.todo.common.ui
 
 import androidx.compose.runtime.Composable
 import com.arkivanov.decompose.extensions.compose.jetbrains.Children
-import example.todo.common.edit.ui.TodoEditContent
-import example.todo.common.main.ui.TodoMainContent
 import example.todo.common.root.TodoRoot
 import example.todo.common.root.TodoRoot.Child
-import example.todo.common.utils.crossfade
 
 @Composable
 fun TodoRootContent(component: TodoRoot) {
