+++ examples/todoapp/common/edit/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     id("multiplatform-setup")
-    id("multiplatform-compose-setup")
+    id("android-setup")
 }
 
 kotlin {
