+++ examples/todoapp/common/utils/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     id("multiplatform-setup")
-    id("multiplatform-compose-setup")
+    id("android-setup")
 }
 
 kotlin {
@@ -9,9 +9,8 @@ kotlin {
             dependencies {
                 implementation(Deps.ArkIvanov.MVIKotlin.rx)
                 implementation(Deps.ArkIvanov.MVIKotlin.mvikotlin)
-                implementation(Deps.ArkIvanov.MVIKotlin.mvikotlinExtensionsReaktive)
                 implementation(Deps.ArkIvanov.Decompose.decompose)
-                implementation(Deps.ArkIvanov.Decompose.extensionsCompose)
+                implementation(Deps.Badoo.Reaktive.reaktive)
             }
         }
     }
